% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMermecApp.R
\name{runMermecApp}
\alias{runMermecApp}
\title{runMermecApp}
\usage{
runMermecApp(forceH2Oriavvio = FALSE)
}
\arguments{
\item{forceH2Oriavvio}{se valore è VERO (TRUE) allora all'avvio viene
forzato il riavvio dell'engine AI di H2O}
}
\description{
Esegue il software, vedi esempio.
}
\examples{
# runMermecApp()
# Scarica il file di esempio cliccando in alto a destra
}
