% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMermecApp.R
\name{runMermecApp}
\alias{runMermecApp}
\title{runMermecApp}
\usage{
runMermecApp(...)
}
\arguments{
\item{...}{parametri di input, ma possono essere lasciati vuoti}
}
\description{
Esegue il software, vedi esempio
}
\examples{
# runMermecApp()
}
